requirements.txt
google-genai
python-dotenv
social-post-api
